Fix shift-overflow warning for CPU_FEATURE_BIT_AVX512VL
Use (int)(1u << 31) instead of 1 << 31 to avoid -Werror=shift-overflow=
when shifting into the sign bit of a signed int. Keeps the constant type
consistent with all other CPU_FEATURE_BIT_* constants.

+++ common/sys/sysinfo.cpp
@@ -248,7 +248,7 @@ namespace embree
   static const int CPU_FEATURE_BIT_AVX512ER = 1 << 27;    // AVX512ER (exponential and reciprocal instructions)
   static const int CPU_FEATURE_BIT_AVX512CD = 1 << 28;    // AVX512CD (conflict detection instructions)
   static const int CPU_FEATURE_BIT_AVX512BW = 1 << 30;    // AVX512BW (byte and word instructions)
-  static const int CPU_FEATURE_BIT_AVX512VL = 1 << 31;    // AVX512VL (vector length extensions)
+  static const int CPU_FEATURE_BIT_AVX512VL = (int)(1u << 31);    // AVX512VL (vector length extensions)
   static const int CPU_FEATURE_BIT_AVX512IFMA = 1 << 21;  // AVX512IFMA (integer fused multiple-add instructions)
   
   /* cpuid[eax=7,ecx=0].ecx */
